removed useless code + reverse array fix
- removed useless legacy construction world code
~ fixed reverse world construction being always 25% off
~ refactored code to be slightly faster
~ changed array annotation from List[List[float]] to proper List[List[int]]

+++ bucket.py
@@ -43,7 +43,7 @@ class Bucket:
                 whites[ind] += macro
         return [count/len(self.photos) for count in whites]
     
-    def get_whites_count_int(self) -> List[float]:
+    def get_whites_count_int(self) -> List[int]:
         whites = [0 for _ in range(self.k)]
         for photo in self.photos:
             for ind, macro in enumerate(photo.photo):
